Try setting PY for release.sh script
For some reason, when building 3rd party wheels $PY is resolving to system python! This is not good, as we need it to use the UCS2 pyenv installed Python. See https://travis-ci.org/pantsbuild/pants/jobs/499114720#L2360.

We already set $PY in the Dockerfile, but maybe it's somehow getting overriden or something? Try setting it here to see if that changes things.