Try setting PY for bootstrap also
Now the bootstrap is not consistently using 2.7.15 (UCS2) :( I'm not sure why. I don't think it's from changing the Travis stage, so my next idea is that it's from setting PY on the right side of the command and not the left - I'm not sure how && works with env vars.