add simple support for rerun-if-(env-)changed
To build always valid binaries, the compile task implements a simple solution to support `cargo:rerun-if-env-changed` and `cargo:rerun-if-changed` statements.
If a build script target contains a `cargo:rerun-if-env-changed` or `cargo:rerun-if-changed` statement, the compile task marks that target and its dependent targets as invalid. The task itself doesn’t actually check if an environment variable or a file has changed.