Properly test that we're building the right UCS
Stu and Danny explained to me that the PEX phrase "Using the current platform" is not necessarily a good indicator for how `./pants.pex` is built. For example, they explained that it shouldn't matter which interpreter Conan Pex uses in this context.

Instead, it's far more robust to explicitly test the generated `./pants.pex` to ensure it works as expected. Specifically, we adopt https://stackoverflow.com/questions/1446347/how-to-find-out-if-python-is-compiled-with-ucs-2-or-ucs-4#1446456.

This is more robust and will help to ensure we don't add regressions.