Revert "add a lint to warn on implicit string concatenation (#7286)" (#7340)
This reverts commit 3a81e86eca33bcd16ca68838052ffeb25fdcc803.

### Problem

The use of the `asttokens` library, now within the pythonstyle checker pex, is causing a variant of #7158 when consumed outside of this repo. Additionally, some files fail to parse with:
```
    for tok in tokenize.generate_tokens(StringIO(str_node_text).readline):
  File "/opt/ee/python/2.7/lib/python2.7/tokenize.py", line 374, in generate_tokens
    ("<tokenize>", lnum, pos, line))
  File "<tokenize>", line 3
    "else echo 'no_op'; fi)"
```

Suppressing the lint only works if the lint runs without error, so this makes it impossible to consume outside of the pants repo right now (or at least, impossible in a lot of repos outside of the pants repo).

### Solution

- Revert #7286.

### Result

This will be unreverted asap when the above issues are investigated and fixed.