Improve release.sh check for valid interpreter
We really want to check two things: 1) that the interpreter is available on the PATH, and 2) that they are using python2.7, as the script does not (yet) work with Python 3.

Rather than muddling both these two tests into one command, instead split them out into distinct checks with very explicit error messages.

This will also make it easier to ensure they are using only 2.7 or 3.6 to release, once that PR goes through, as we don't want to allow <2.7, 3.0<=x<=3.5, or 3.7.

Thanks Benjy!