Fix Docker not evaluating external env var
THe bootstrap shards were failing to read their args (`-b` vs. `-2b`). This is because Docker expects a string with its exact arguments, and will not do any env var resolution from the external environment. Instead, all env vars must be valuated by the parent bash process and then the result passed to Docker.

This leads to slightly more duplication, but in some ways is more explicit and is an acceptable tradeoff.