Provide comments explaining how to use env vars with templates
Danny pointed out it can be a bit confusing that we have two levels of templating: the mustache template and the env vars that it populates.

Add comments to the mustache templates stressing that it is expecting env vars and explaining what it expects these to be. This is kinda like docstring for mustache.

Also add comments in .travis.yml where we have a base image define a script, but then expect the callers of the anchor to provide env vars. Make this expectation clear.