Move comments directly into help argument
Danny and Stu have a point that it's better to expose this info directly to the user, rather than keeping it in a comment in source code.

I originally was thinking we'd want to keep the help message shorter, and rely on the message in the runtime check to steer people the right way, but this is bad reasoning. I can't think of a time I was upset a help message was too long (so long as it was concise and everything had a reason to be there)..