Convert integration test to 2 unit tests
Integration test was never going to work because we run from pants.pex and can't change the interpreter for the subprocess. Even if we could, the subprocess would start bootstrapping Pants, which we do NOT want.

Further, patching the integration test won't work because the ./pants run is a subprocess so won't pick up on the patch.

Instead, we can split the functionality into two dedicated unit tests and actually end up with better coverage and faster tests!