Refactor test code's interpreter_selection_utils.py for better naming of util functions (#7366)
### Problem
We currently have multiple functions named `skip_unless_pythonX` which skip unless the Python version is present on the system. The docstring explains this.

However, the name suggests it will skip if the current interpreter itself is not python27, rather than the version being available to begin with. This could be made clearer by updating the name.

### Solution
Rename `skip_unless_pythonX` to `skip_unless_pythonX_present`. This is safe to do without a deprecation because this is internal test util code, and not exposed as an API.

Also refactor below changes:

* Define `PY_X` for more Python versions for better consistency.
* Add `skip_unless_all_pythons_present()`.
   * We have a function to check `all`, and it's helpful to check `any`.
   * For example, we might want to run the test if either Py36 or Py37 are available, and we don't care that both are.
* Add `find_all_pythons_present()`.
   * This is useful when paired with `skip_unless_all_pythons_present()`. If the test is not skipped because at least one requested version is available, this function may be used to identify which are valid.