Add runtime check for valid Python interpreter (#7365)
### Problem
Now that we'll soon allow running Pants with more than Python 2.7 for every day users, we should eagerly ensure that they are using a supported Python version. This is necessary because Pants is typically ran as a script, rather than as a library, and the script may be ran with whichever interpreter the user chooses.

Even though the setup repo is supposed to ensure people use a valid interpreter, it is easy to get around this with `PYTHON=3.5 ./pants` or with putting `pants_engine_python_version: 3.5` in their `pants.ini`. Both these scenarios are totally valid to do, and we should not teach the setup repo or `--pants-engine-python-version` what are valid options, because the former is easy to fall out of sync with users and the latter is easy to circumvent.

While Pants _might_ work with something like 3.5, we should at least warn users they're treading into dangerous waters. If they want to proceed by setting an env var bypass, all power to them. But for the sake of UX we should at least give the warning, rather than relying on strange bugs and `SyntaxError`s to enforce using the correct version.

### Solution
Add a runtime check to `pants_loader.py` with a human readable error message. Allow skipping the check with `PANTS_IGNORE_UNSUPPORTED_PYTHON_INTERPRETER`.

Add unit tests for this all.

### Result
When running with anything other than 2.7 or 3.6+, Pants will error out.

Once we drop Py2, we can change this check to 3.6+.