Use UUID from run_id as trace_id if trace_id isn't set by a flag (#7319)
Every Pants run has an ID (run_id, punts_run_...) and last part of this ID is a 32 characters hex string.
trace id in Zipkin traces is a 16 or 32 characters hex string. That means that Pants run_id can be used as a Zipkin trace id.
That may make it easier to correlate Zipkin traces with other places that we already log a run_id.