Fix rule graph nondeterminism due to undetected ambiguity (#7379)
### Problem

Integration tests of `@console_rules` would frequently flake (as in #6782) with rules self-cycling in a peculiar way. Investigating the failure lead to the conclusion that the `RuleGraph` was being constructed non-deterministically.

The source of the non-determinism was that we were "simplifying" all instances of one `@rule` that used the same number of `Params`, and treating them interchangeably. But that simplification was subject to rust's randomized hash ordering, and so we would choose which of the `Params` to use at random.

After fixing that non-determinism, we also needed to solve the underlying ambiguities (there were two or three) that lead to there being multiple options to choose from.

### Solution

* Change `GraphMaker::choose_dependency` to not simplify rules, and instead preserve all candidates with the current minimum `Param` set size.
* Fix the ambiguity that lead to multiple choices being available by introducing an (intuitive, I think) restriction that only subgraphs that actually use the `Param` provided by a `Get` are eligible to satisfy it. See the explanation in the test.

### Result

Unskips `@console_rule` integration tests, adds a (previously flaky, now stable) unit test of the ambiguous case, and fixes a broken test detected by the new restriction on rule graph parameters. Fixes #6782.