Set PEX_PYTHON_PATH in v2 python test running
Without this, the system's default python interpreter will be used
(possibly via pexrc).

Long term, we should probably be using a flag to specify what version of
python should be used for running tests, but in the short term, we're
currently using Pants's python interpreter to make the pex, so we should
_also_ be using it to run the pex, otherwise we get errors finding
dependencies for the correct platform.