[rsc-compile] support explicit tagging of targets in rsc compile (#7362)
In order to be able to incrementally enable rsc-then-zinc, we need to be able to opt targets into rsc, and to be able to opt them out if there are issues. This adds tags for opting targets into either rsc or zinc.
    
It also causes setting the default workflow on the CLI to override target tag values. This allows for testing workflows on targets that are tagged with the other workflow.