Collect Zipkin spans for v2 engine (#7342)
### Problem
Currently, the v2 engine is a black box when it comes to collecting [Zipkin spans ](https://zipkin.io/pages/instrumenting.html)for tracing.

### Solution

This PR allows tracking spans for v2 engine for every Node. But, the number of these Nodes is very big (hundreds or thousands depending on the size of the target), so, later on, will be decided for what Nodes Zipkin spans will be collected. So a flag was added that enables Zipkin tracing for v2 engine.

### Result
There are the next things in this PR:
- workunits for v2 engine that collect Zipkin tracing info;
- flag that enables collecting Zipkin spans for v2 engine.