Allow reporters to see the correct end_time for a workunit (#7389)
Problem:
The end_time for a workunit is not accessible to reports, since it's set after we tell the report that the workunit is complete.

Solution:
Set the end_time on the workunit before notifying reporter's that it's complete.

Result:
Reporters can now access an accurate end_time for workunits. This will be used for adjacent work (in flight) that returns a more complete representation of the build than is currently returned by the RunTracker/reporters.