Release pants as both a Python 2.7 PEX and a Python 3.6 PEX (#7401)
### Problem
While wheels are now released with Python 3 support thanks to https://github.com/pantsbuild/pants/pull/7197, this is only one of two ways people consume Pants. The other way is through a Pex released to Github. 

We decided to release both a Py27 and Py36 Pex. 

We will not be releasing a Py37 Pex, as it is would require Py37 wheel building shards which would require a Centos7 docker image. We decided this is not worth the work for how few people we anticipate would use Py37 with Pex combined.

We will also not be releasing a more flexible Pex that works with either Py27 or Py36, as it is blocked by https://github.com/pantsbuild/pex/issues/654. We also want organizations pinning their Python version, so a flexible release is not particularly useful in this instance.

This implements https://github.com/pantsbuild/pants/issues/6450#issuecomment-458686132.

### Solution
Release both `pants.${VERSION}.py{27,36}.pex`.

This requires changing how we run `release.sh -p` and adding new CI shards to deploy on both Py2 and Py3.

### Result
The next release, we will release both `pants.1.15.0.rc1.py27.pex` and  `pants.1.15.0.rc1.py36.pex`.