Document pantsd startup and Pailgun behaviours (#7376)
### Problem
The codepaths that executing pants with --enable-pantsd are hard to understand. While there are good in-place comments, there is not an end-to-end high level description of the components in play and their interaction.

### Solution
Write up that documentation. Specifically, this addition aims to explain what happens when you run ./pants --enable-pantsd list src/scala:: while pantsd is off, and give a high-level overview of the components necessary to understand that codepath.

### Result
Documentation changes in pantsbuild.org, which contain:

  Descriptions of ProcessManager, PantsDaemon, PantsService, the Pailgun protocol and the classes related to its functioning.
  An end-to-end example with links to the code.
This is reviewable (because what's there is unlikely to change), but there is work to be done before this is mergeable:

  Add introduction paragraph which describes the high-level goal of Pantsd.
  Add links to the repo for the classes mentioned here. We may not need link to every method, but links to every class would be nice.
  Integrate with the pantsbuild.org website, possibly as a tab on the left.

Closes #5114
Related to #6069, possibly closing it