Add JSON Reporter for more detailed workunit stats (#7392)
### Problem

Currently there is no single source of detailed structured data for builds. `RunTracker` has the ability to post certain stats to external URLs or a json on the user's machine, but these stats do not include much information of each workunit. `HtmlReporter` has more detail, but not structured in a way that could be easily consumed by others.

Creating such a single source for build data could be leveraged in different use cases:
- More data that the `RunTracker` could post externally (that would survive a `clean-all`)
  - If this supported sending deltas, the build could be viewed remotely in near-real time
- Improve other reporters by having a single json that could be polled for deltas / progress

### Solution

This adds a new `JsonReporter` that collects the outputs from each workunit and writes them to (currently) a json file (similar to how the `HtmlReporter` writes outputs to an `html` dir).

There are several more steps to plug this in fully for reporting these stats to an external URL, though that work could be broken up over subsequent PRs, specifically:
- How to incorporate a `JsonReporter` into the current stats that are reported (perhaps by a versioning flag on the run tracker - which leads to the question of how tightly coupled to the `RunTracker` this should be).

Additionally, things that aren't solved here, but could be done as next steps:
- Compressing the json for a completed build, since it's conceivable that those reports could be quite large.
- Sending deltas during the build instead of a complete file at the end.