Refactor tests to not use a setUp function
Now that we can use self.execute_task(), the only value we were pinning is temp_pants_ini_path. Once we change Pants to allow running without a pants.ini file created, we won't want this always created for the test case of a missing pants.ini. So, this is sort of pre-work + it leads to less code.