nailgun connect timeout error message fix (#7437)
### Problem

We previously had an untested code path (`NailgunExecutor._await_socket()`) which constructed a `NailgunClient.NailgunError` upon a timeout when connecting to a started nailgun subprocess. The `NailgunError` constructor was changed a while ago to create its own error message, and instead accept arguments such as the pid and pgrp in its constructor, but this untested code path hadn't been updated to provide those arguments. This caused a strange error when this unlikely timeout occurred while compiling zinc in an osx laptop environment.

### Solution

- Raise a `self.InitialNailgunConnectTimedOut` and convert it into a `self.Error` in a try/except instead of raising a `NailgunError`.
- Add testing for nailgun connection timeouts.
- Create a separate option `--nailgun-subprocess-startup-timeout` to cover the time to wait for the nailgun subprocess to start up, using `--nailgun-timeout-seconds` to cover the time to wait after that for the nailgun subprocess to begin producing output.

### Result

Another code path in our nailgun interactions is tested, and a bug in that code path is fixed.