Can't use the folder /usr/local/aws for cache with Linux
Travis runs into a permissions issue when doing so: https://travis-ci.org/pantsbuild/pants/jobs/513927894#L432.

Further, the only reason we were doing this is the tutorial used this install dir.

Instead, use a custom install dir. Our symlink makes it so that it doesn't really matter where we install it.