No longer pin `pants_runtime_python_version` with `./pants generate-pants-ini` (#7469)
In https://github.com/pantsbuild/setup/pull/49, we're changing the setup repo's `./pants` script to default to Python 3.6, and then fall back to Python 2.7. Thanks to that, we no longer need to use `pants_runtime_python_version` to get the script to use Python 3.

In fact, we should not set `pants_runtime_python_version`. The Python version used is an implementation detail we don't want users to have to worry about, and we will be deprecating the option in 1.17.0.dev0. So, us auto-generating the option when it is not even necessary is asking for confusion and a worse experience for users.