No longer use a custom pyenv_root
If there are already pyenv installed Pythons, then we want to be able to use them. Namely, we're going to be porting some of our build scripts to Python 3 and we need to ensure we have Python 3 on every shard. Overriding it means we could not use the Python 3 installed on centos6.

We can point to ~/.pyenv/versions/2.7.15/bin/python to disambiguate.