Check if PYENV_ROOT exists, rather than PYENV_BIN
In normal Travis, we have to check if PYENV_BIN exists because Travis will automatically create the TRAVIS_ROOT folder as a part of caching.

Docker, however, does not have caching, so we can simply check if the root folder exists.

If we were to stick with PYENV_BIN, we would need to `rm -rf PYENV_ROOT` to make sure the git clone would not fail which is more work than we want.