Remove StatsDb / Timing stats views (#7475)
Followup to #7446.

The schema for StatsDB is coupled to the current output of stats used in the RunTracker - which complicates any attempts to move to a new stats schema (ie. the JsonReporter).

Given that the only usage of StatsDB was for the timing views from ./pants server, and those views are not heavily used (as far as could be determined in related Slack convo), this should be safe to remove.