Remove @rule Select syntax (#7477)
### Problem

The `Select` syntax in `@rule` declarations used to differentiate the declarative parameter selectors, but with the advent of coroutine `@rules` and removal of `TypeConstraint`, the `Select` syntax no longer has any use.

### Solution

Remove the `Select` syntax for `@rules`, while preserving it as the marker for root selections in the graph. Update the documentation for `@rules` to refer to these as "parameter selectors".