Move JsonReporter setup to RunTracker (when v2 stats is configured)
This allows us to only run json reports when v2 stats are specified, and since the JsonReporter is _only_ used by the RunTracker for stats it feels like a better fit here than in Reporting / Report. Additionally, we can leverage the in-memory results of the json reporter (instead of reading the file) and more clearly illustrate the differences between the v1 / v2 stats when selecting them in the RunTracker based on the configured flag.