Stop reporting the 3 slowest tests
This added about 5 lines of output to every single test file we run.

There is a time and place to audit how long tests take, but 98% of the time this just adds noise to our daily CI runs. If we want to audit test time, we can simply run a dedicated CI run with this option enabled.