Invalidate dependency targets for jvmdoc if --combined. (#7455)
### Problem

The doc.scaladoc task (and likely also the doc.javadoc task), when used with --combined is skipping dependencies in the graph during the 'invalidation' stage, because it is not passing invalidate_dependents in the docable predicate that is part of the base class JvmDocGen.

See issue: #7434

### Solution

In the relevant spot in `JvmDocGen` class, this change passes the `invalidate_dependents` parameter to `self.invalidate()` with attribute `self.combined` for the value which is `True` for combined doc builds and `False` otherwise. 

### Result

This causes the invalidation logic to behave as expected, specifically, invalidating the dependencies within the entire graph for a combined doc build. I tested this locally to ensure the correct behavior, and ran all relevant unit tests.