remove pyenv bootstrapping from travis_ci/Dockerfile (#7505)
### Problem

The conclusion of #7064 was, after `pantsbuild/centos6:latest` was updated, to remove the pyenv bootstrapping in `travis_ci/Dockerfile` as it was now being done in the centos6 base image.

### Solution

- Remove the marked section.

### Result

This will speed up all Docker shards that had Py3 installed by 2-3 minutes. So, Linux build engine Py36 and Linux build wheels abi3.

This means overall CI is about 5 minutes faster, and the bootstrap stage 2-3 minutes faster, which is especially significant since that blocks starting other tests.

Finally, this means everywhere we run Pants in CI, we have Python 3 installed now (instead of having to install it again!).