make pantsd signal integration tests more reliable by setting the timeout (#7504)
### Problem

An attempt to resolve #7425. As mentioned on that ticket, the default (reasonable) timeout for waiting for pailgun subprocesses to complete after a signal (`--pantsd-pailgun-quit-timeout`) was 1 second. I believe the remaining issue here is that some invocations in CI would actually take more than a second to quit, which would cause nondeterministic behavior.

### Solution

- Set the timeout to 5 seconds on any test which intends for its subprocess to complete within the timeout.
- Match more of the stderr contents in each error case to remove ambiguity.

### Result

`TestPantsDaemonIntegration.test_pantsd_sigterm` hopefully won't continue to flake!!