pass through an RpcFinished(None) when grpc write streams are terminated early (#7422)
### Problem

When using remotely executed scala compiles by setting `--execution-strategy=hermetic` with zinc compiles and pointing the `--remote-execution-server` and `--remote-store-server`s at an internal instance of https://github.com/twitter/scoot to compile scala in https://github.com/twitter/util, we are seeing some [`RpcFinished(None)` errors](https://docs.rs/grpcio/0.4.2/grpcio/enum.Error.html) causing upload digest requests to fail after a few retries.

In #6344 we see a case of a spurious `RpcFinished(None)` which caused us to fork grpcio. We believe this is similarly a spurious failure caused by a non-erroneous race condition, as in logs from Scoot we are seeing that the digests reporting a failure have successfully been uploaded to the remote store. In testing, we have found that this occurs with files which are large and are used by multiple process requests (e.g. `scala-reflect.jar` in zinc compilations). We believe the `RpcFinished(None)` we are converting into an `Ok(None)` here occurs when pants has multiple concurrent write requests against the remote store, and the remote store cancels the others after it successfully receives the first write request.

### Solution

- Pass through an [`RpcFinished(None)`](https://docs.rs/grpcio/0.4.2/grpcio/enum.Error.html) in `store.rs`.

### Result

Pants no longer considers it a failure to receive an `RpcFinished(None)` response. The discussion is ongoing as to whether this should be an accepted part of the remote execution API at https://groups.google.com/forum/#!topic/remote-execution-apis/NXUe3ItCw68/discussion.