Enable pantsd in Travis (#7440)
** Problem **
See #7320 for context. We would like to run integration tests under pantsd, to discover more issues.

** Solution **
Enable pantsd in the integration tests in the cron shard for py36 (not in the test runner, but on individual integration tests themselves) via a class property, and blacklist the tests that don't work with a useful docstring.

** Result **
Pantsd will now be enabled in integration and contrib tests running in the cron py36 shards.